+++ scheduler/client.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: client.c,v 1.91.2.21 2002/09/25 12:52:41 mike Exp $"
+ * "$Id: client.c,v 1.91.2.22 2002/09/26 15:19:50 mike Exp $"
  *
  *   Client routines for the Common UNIX Printing System (CUPS) scheduler.
  *
@@ -1598,7 +1598,8 @@ SendError(client_t      *con,	/* I - Connection */
 
     if (httpPrintf(HTTP(con), "Content-Type: text/html\r\n") < 0)
       return (0);
-    if (httpPrintf(HTTP(con), "Content-Length: %d\r\n", strlen(message)) < 0)
+    if (httpPrintf(HTTP(con), "Content-Length: %d\r\n",
+                   (int)strlen(message)) < 0)
       return (0);
     if (httpPrintf(HTTP(con), "\r\n") < 0)
       return (0);
@@ -2624,5 +2625,5 @@ pipe_command(client_t *con,		/* I - Client connection */
 
 
 /*
- * End of "$Id: client.c,v 1.91.2.21 2002/09/25 12:52:41 mike Exp $".
+ * End of "$Id: client.c,v 1.91.2.22 2002/09/26 15:19:50 mike Exp $".
  */
